# openapi/products_spec.yaml
openapi: 3.0.1
info:
  description: Products API
  title: Products API
  version: v1.0.0
servers:
- url: http://products_url
paths:
...
  /v1/api/payments/redact-reference/{govukPaymentId}:
    post:
      operationId: redactReference
      parameters:
      - in: path
        name: govukPaymentId
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            '*/*': {}
          description: default response
      tags:
      - Payments
  /v1/api/payments/{gatewayAccountId}/{referenceNumber}:
    get:
      operationId: findPaymentsByGatewayAccountIdAndReferenceNumber
      parameters:
      - example: 1
        in: path
        name: gatewayAccountId
        required: true
        schema:
          type: integer
          format: int32
      - example: RE4R2A6VAP
        in: path
        name: referenceNumber
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: OK
        "404":
          description: Not found
      summary: Find payment by gateway account ID and reference number.
      tags:
      - Payments
  /v1/api/payments/{paymentExternalId}:
    get:
      operationId: findPaymentByExternalId
      parameters:
      - example: h6347634cwb67wii7b6ciueroytw
        in: path
        name: paymentExternalId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: OK
        "404":
          description: Not found
      summary: Find payment by payment external ID
      tags:
      - Payments
...
  /v1/api/products/{productExternalId}/payments:
    get:
      operationId: findPaymentsByProductExternalId
      parameters:
      - example: uier837y735n837475y3847534
        in: path
        name: productExternalId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
          description: OK
        "404":
          description: Not found
      summary: Find list of payments that belongs to the specified product external
        ID.
      tags:
      - Payments
    post:
      operationId: createPayment
      parameters:
      - example: uier837y735n837475y3847534
        in: path
        name: productExternalId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: number
                  description: Price override for the payment amount. If not present
                    this will defaults to price of product.
                  example: 9090
                  maximum: 10000000
                  minimum: 1
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
          description: OK
        "400":
          description: For invalid payload
      summary: Creates new payment for a given product
      tags:
      - Payments
...
components:
  schemas:
...
    Link:
      type: object
      properties:
        href:
          type: string
          example: http://products-url/v1/api/products/874h5c87834659q345698495
        method:
          type: string
          example: GET
        rel:
          type: string
          enum:
          - self
          - pay
          - next
          - friendly
          example: self
    Payment:
      type: object
      properties:
        _links:
          type: array
          example:
          - href: https://govukpay-products.cloudapps.digital/v1/api/payments/h6347634cwb67wii7b6ciueroytw
            method: GET
            rel: self
          - href: https://some.valid.url/paid
            method: GET
            rel: next
          items:
            $ref: '#/components/schemas/Link'
        amount:
          type: integer
          format: int64
          example: 1050
        external_id:
          type: string
          example: h6347634cwb67wii7b6ciueroytw
        govuk_payment_id:
          type: string
          example: 7cs487heites2nne5k17j5j9as
        govuk_status:
          type: string
          example: success
        product_external_id:
          type: string
          example: uier837y735n837475y3847534
        reference_number:
          type: string
          example: RE4R2A6VAP
        status:
          type: string
          enum:
          - CREATED
          - SUBMITTED
          - ERROR
          example: SUBMITTED
      required:
      - _links
      - amount
      - external_id
      - product_external_id
      - status
...